Add scripts for the next song, status, and startup
I added separate scripts for going to the next song, getting the
status, and starting the VLC player.

The next song was actually the first one I created, which I did a day
or so ago. I decided to add the start-up next, since that naturally
needs to come first. And once I was done with that, it was really easy
to spin off that code into a status command.

Y'know, I originally envisioned a bunch of separate files. But I'll be
honest - I'm already seeing a lot of code reuse potential. In fact,
call me crazy, but I think this might go better as one giant file with
a whole heap of functions...