Condense the three scripts into a single script
I admit that it's going to be a long file, but merging all of our
scripts into a single file just makes a lot of sense. This has already
allowed us to reuse a LOT of code. We have generic property
get-and-set functions, which we can then call to get status, which
makes launching the program easy.

Now I just need to work on a way for the user to select which command
they want to do...