Execute commands as they are given to us
What I really wanted for this script was something like git - with
it's "git status", "git commit", etc. I wanted to be able to do
"dbus-vlc volume 1.0", or something like that.

After a lot of searching, I found some code to help me do that. It
even allows for some of the commands to have arguments. Perfect!
As an added feature, the way the code is set up, it will execute all
commands in the sequence given. It's a bit excessive honestly, but I
dig that.