Flesh out most of our commands
After a lot of guess-and-check, I've managed to fill out most of our
commands. Did you know that dbus gives an "UnknownMethod" error if
one of the arguments to your methods is wrong? Yeah, it just gives
that error a lot. Can be pretty hard to sort through.

We've got a pretty basic pallette of commands going right now. The
clear command is complex enough that I think I'll need some
development time for that.

I also kinda want to expand the command pallette to add value retrieve
functions but... it's not a priority for me. This is still a personal
project, after all.