Complete the clear command
When I switch between playlists, I typically need to clear away all of
the tracks that were there previously. A clean slate. That's what the
clear command is for.

This is actually the most complex of all the commands I've developed.
We have to parse through the output given by a property Get. That
actually ended up being pretty easy, though with bash's syntax
anything can be made a mess.

There's one last thing I want to do - get commands for volume and
rate.