Add get functionality for volume and rate
I could see instances where we would want to set the volume and the
rate contextually - i.e. rather than setting the rate or volume to
an arbitrary value, we want to set the volume or rate to something
plus-or-minus it's current value. To do that, we need to get the
value.

So I added rate and volume get commands. They just print the current
values to stdout. But that's all we'll need to do what we want.